refactored methods out to utility
